Deleting of webhooks added. Some files moved to packeges for better structure.
