+++ src/main/java/com/mattermost/integration/figma/input/oauth/RemoteOauth2.java
@@ -1,4 +1,4 @@
-package com.mattermost.integration.figma.input;
+package com.mattermost.integration.figma.input.oauth;
 
 import lombok.Data;
