+++ src/main/java/com/mattermost/integration/figma/notification/service/FileNotificationService.java
@@ -3,21 +3,23 @@ package com.mattermost.integration.figma.notification.service;
 import com.mattermost.integration.figma.input.file.notification.FileCommentNotificationRequest;
 import com.mattermost.integration.figma.provider.FigmaTokenProvider;
 import com.mattermost.integration.figma.provider.NgrokLinkProvider;
-import com.mattermost.integration.figma.security.dto.OAuthCredsDTO;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpMethod;
 import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.util.LinkedMultiValueMap;
+import org.springframework.util.MultiValueMap;
 import org.springframework.web.client.RestTemplate;
 
 @Service
 public class FileNotificationService {
-    private static final String FILE_NOTIFICATION_URL = "https://api.figma.com/v2/webhooks";
+    private static final String BASE_WEBHOOK_URL = "https://api.figma.com/v2/webhooks";
     private static final String PASSCODE = "Mattermost";
     private static final String FILE_COMMENT_EVENT_TYPE = "FILE_COMMENT";
-    private static final String REDIRECT_URL = NgrokLinkProvider.REDIRECT_URL;
-    private static final String WEBHOOK_URL = "/fileComment";
+    //Production Mattermost link
+    private static final String REDIRECT_URL = "http://localhost:8066/plugins/com.mattermost.apps/apps/spring-boot-example%s?secret=%s";
+    private static final String FILE_COMMENT_URL = "/webhook/comment";
 
     private final RestTemplate restTemplate;
 
@@ -25,7 +27,7 @@ public class FileNotificationService {
         this.restTemplate = restTemplate;
     }
 
-    public String subscribeToFileNotification(String teamId) {
+    public String subscribeToFileNotification(String teamId, String webhookSecret) {
         if (teamId != null && !teamId.isEmpty() && !teamId.isBlank()) {
             HttpHeaders headers = new HttpHeaders();
             headers.setContentType(MediaType.APPLICATION_JSON);
@@ -34,12 +36,23 @@ public class FileNotificationService {
             fileCommentNotificationRequest.setEventType(FILE_COMMENT_EVENT_TYPE);
             fileCommentNotificationRequest.setTeamId(teamId);
             fileCommentNotificationRequest.setPasscode(PASSCODE);
-            fileCommentNotificationRequest.setEndpoint(REDIRECT_URL.concat(WEBHOOK_URL));
+            //For production Mattermost link
+            //fileCommentNotificationRequest.setEndpoint(String.format(REDIRECT_URL, FILE_COMMENT_URL, webhookSecret));
+            fileCommentNotificationRequest.setEndpoint(NgrokLinkProvider.REDIRECT_URL.concat(FILE_COMMENT_URL));
             System.out.println(fileCommentNotificationRequest);
             HttpEntity<FileCommentNotificationRequest> request = new HttpEntity<>(fileCommentNotificationRequest, headers);
-            restTemplate.postForEntity(FILE_NOTIFICATION_URL, request, String.class);
+            restTemplate.postForEntity(BASE_WEBHOOK_URL, request, String.class);
             return "{\"text\" : \"Success\"}";
         }
         return "{\"text\" : \"There is no such team id\"}";
+    }
+
+    public void deleteWebhook(String webhookId) {
+        MultiValueMap<String, String> headers = new LinkedMultiValueMap<>();
+        headers.add("Authorization", String.format("Bearer %s", FigmaTokenProvider.token.getAccessToken()));
+        HttpEntity<Object> request = new HttpEntity<>(headers);
+        String url = BASE_WEBHOOK_URL.concat("/").concat(webhookId);
+        restTemplate.exchange(url, HttpMethod.DELETE, request, String.class);
+        System.out.println("Successfully deleted webhook with id: " + webhookId);
     }
 }
