+++ src/main/java/com/mattermost/integration/figma/provider/NgrokLinkProvider.java
@@ -2,5 +2,5 @@ package com.mattermost.integration.figma.provider;
 
 //Temp class to hold local Ngrok link
 public class NgrokLinkProvider {
-    public static final String REDIRECT_URL = "https://1877-213-109-232-43.ngrok.io";
+    public static final String REDIRECT_URL = "https://6a01-213-109-232-169.ngrok.io";
 }
