+++ src/main/java/com/mattermost/integration/figma/webhook/WebhookController.java
@@ -1,7 +1,8 @@
 package com.mattermost.integration.figma.webhook;
 
 
-import com.mattermost.integration.figma.input.InputPayload;
+import com.mattermost.integration.figma.input.oauth.InputPayload;
+import com.mattermost.integration.figma.input.file.notification.FileCommentWebhookResponse;
 import com.mattermost.integration.figma.notification.service.FileNotificationService;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -19,13 +20,17 @@ public class WebhookController {
     }
 
     @PostMapping("/comment")
-    public void comment(@RequestBody String request) {
-        System.out.println("Got it "+ request);
+    public void comment(@RequestBody FileCommentWebhookResponse response) {
+        System.out.println(response);
+        if (!response.getEventType().equals("PING")) {
+            fileNotificationService.deleteWebhook(response.getWebhookId());
+        }
     }
 
     @PostMapping("/subscribe/fileComment")
     public String subscribeToFileComment(@RequestBody InputPayload request) {
         System.out.println(request);
-        return fileNotificationService.subscribeToFileNotification(request.getValues().getTeamId());
+        return fileNotificationService.subscribeToFileNotification(request.getValues().getTeamId(),
+                request.getContext().getApp().getWebhookSecret());
     }
 }
